services/visr_api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9